@@ -1,5 +1,5 @@
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 show_column_numbers = true
 show_error_context = true
 ignore_missing_imports = true
@@ -42,7 +42,7 @@ disallow_untyped_calls = false
 
 [tool.ruff]
 line-length = 79
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 # enable the following rule classes:
