@@ -10,7 +10,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import typing as ty
 
 __all__ = ('ClientException',)
 
@@ -20,6 +19,6 @@ class ClientException(Exception):
 
     message = "ClientException"
 
-    def __init__(self, message: ty.Optional[str] = None):
+    def __init__(self, message: str | None = None):
         self.message = message or self.message
         super().__init__(self.message)
