@@ -33,7 +33,7 @@ class _Saml2TokenAuthMethod(v3.AuthMethod):
         auth: v3.Auth,
         headers: dict[str, str],
         request_kwargs: dict[str, object],
-    ) -> ty.Union[tuple[None, None], tuple[str, ty.Mapping[str, object]]]:
+    ) -> tuple[None, None] | tuple[str, ty.Mapping[str, object]]:
         raise exceptions.HttpNotImplemented(
             'This method should never be called'
         )
@@ -57,14 +57,14 @@ class BaseSAMLPlugin(v3.FederationBaseAuth):
         password: str,
         protocol: str,
         *,
-        trust_id: ty.Optional[str] = None,
-        system_scope: ty.Optional[str] = None,
-        domain_id: ty.Optional[str] = None,
-        domain_name: ty.Optional[str] = None,
-        project_id: ty.Optional[str] = None,
-        project_name: ty.Optional[str] = None,
-        project_domain_id: ty.Optional[str] = None,
-        project_domain_name: ty.Optional[str] = None,
+        trust_id: str | None = None,
+        system_scope: str | None = None,
+        domain_id: str | None = None,
+        domain_name: str | None = None,
+        project_id: str | None = None,
+        project_name: str | None = None,
+        project_domain_id: str | None = None,
+        project_domain_name: str | None = None,
         reauthenticate: bool = True,
         include_catalog: bool = True,
     ) -> None:
@@ -125,7 +125,7 @@ class BaseSAMLPlugin(v3.FederationBaseAuth):
 
     @staticmethod
     def str_to_xml(
-        content: bytes, msg: ty.Optional[str] = None, include_exc: bool = True
+        content: bytes, msg: str | None = None, include_exc: bool = True
     ) -> etree._Element:
         try:
             return etree.XML(content)
