@@ -10,7 +10,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import typing as ty
 
 from keystoneauth1 import discover
 from keystoneauth1.identity.generic import base
@@ -32,24 +31,24 @@ class Password(base.BaseGenericPlugin):
     def __init__(
         self,
         auth_url: str,
-        username: ty.Optional[str] = None,
-        user_id: ty.Optional[str] = None,
-        password: ty.Optional[str] = None,
-        user_domain_id: ty.Optional[str] = None,
-        user_domain_name: ty.Optional[str] = None,
+        username: str | None = None,
+        user_id: str | None = None,
+        password: str | None = None,
+        user_domain_id: str | None = None,
+        user_domain_name: str | None = None,
         *,
-        tenant_id: ty.Optional[str] = None,
-        tenant_name: ty.Optional[str] = None,
-        project_id: ty.Optional[str] = None,
-        project_name: ty.Optional[str] = None,
-        project_domain_id: ty.Optional[str] = None,
-        project_domain_name: ty.Optional[str] = None,
-        domain_id: ty.Optional[str] = None,
-        domain_name: ty.Optional[str] = None,
-        system_scope: ty.Optional[str] = None,
-        trust_id: ty.Optional[str] = None,
-        default_domain_id: ty.Optional[str] = None,
-        default_domain_name: ty.Optional[str] = None,
+        tenant_id: str | None = None,
+        tenant_name: str | None = None,
+        project_id: str | None = None,
+        project_name: str | None = None,
+        project_domain_id: str | None = None,
+        project_domain_name: str | None = None,
+        domain_id: str | None = None,
+        domain_name: str | None = None,
+        system_scope: str | None = None,
+        trust_id: str | None = None,
+        default_domain_id: str | None = None,
+        default_domain_name: str | None = None,
         reauthenticate: bool = True,
     ):
         super().__init__(
@@ -80,8 +79,8 @@ class Password(base.BaseGenericPlugin):
         session: ks_session.Session,
         version: discover._PARSED_VERSION_T,
         url: str,
-        raw_status: ty.Optional[str] = None,
-    ) -> ty.Union[None, v2.Password, v3.Password]:
+        raw_status: str | None = None,
+    ) -> None | v2.Password | v3.Password:
         if discover.version_match((2,), version):
             if self._user_domain_id or self._user_domain_name:
                 # TODO(stephenfin): Shouldn't this be an error?
@@ -135,7 +134,7 @@ class Password(base.BaseGenericPlugin):
         return None
 
     @property
-    def user_domain_id(self) -> ty.Optional[str]:
+    def user_domain_id(self) -> str | None:
         return self._user_domain_id or self._default_domain_id
 
     @user_domain_id.setter
@@ -143,14 +142,14 @@ class Password(base.BaseGenericPlugin):
         self._user_domain_id = value
 
     @property
-    def user_domain_name(self) -> ty.Optional[str]:
+    def user_domain_name(self) -> str | None:
         return self._user_domain_name or self._default_domain_name
 
     @user_domain_name.setter
     def user_domain_name(self, value: str) -> None:
         self._user_domain_name = value
 
-    def get_cache_id_elements(self) -> dict[str, ty.Optional[str]]:
+    def get_cache_id_elements(self) -> dict[str, str | None]:
         return {
             'auth_url': self.auth_url,
             'project_id': self._project_id,
