@@ -52,7 +52,7 @@ class OAuth2ClientCredentialMethod(base.AuthMethod):
         auth: base.Auth,
         headers: dict[str, str],
         request_kwargs: dict[str, object],
-    ) -> ty.Union[tuple[None, None], tuple[str, ty.Mapping[str, object]]]:
+    ) -> tuple[None, None] | tuple[str, ty.Mapping[str, object]]:
         """Return the authentication section of an auth plugin.
 
         :param session: The communication session.
@@ -70,7 +70,7 @@ class OAuth2ClientCredentialMethod(base.AuthMethod):
         }
         return 'application_credential', auth_data
 
-    def get_cache_id_elements(self) -> dict[str, ty.Optional[str]]:
+    def get_cache_id_elements(self) -> dict[str, str | None]:
         """Get the elements for this auth method that make it unique.
 
         These elements will be used as part of the
@@ -109,14 +109,14 @@ class OAuth2ClientCredential(base.Auth):
         oauth2_client_id: str,
         oauth2_client_secret: str,
         *,
-        trust_id: ty.Optional[str] = None,
-        system_scope: ty.Optional[str] = None,
-        domain_id: ty.Optional[str] = None,
-        domain_name: ty.Optional[str] = None,
-        project_id: ty.Optional[str] = None,
-        project_name: ty.Optional[str] = None,
-        project_domain_id: ty.Optional[str] = None,
-        project_domain_name: ty.Optional[str] = None,
+        trust_id: str | None = None,
+        system_scope: str | None = None,
+        domain_id: str | None = None,
+        domain_name: str | None = None,
+        project_id: str | None = None,
+        project_name: str | None = None,
+        project_domain_id: str | None = None,
+        project_domain_name: str | None = None,
         reauthenticate: bool = True,
         include_catalog: bool = True,
     ) -> None:
@@ -145,7 +145,7 @@ class OAuth2ClientCredential(base.Auth):
 
     def get_headers(
         self, session: 'ks_session.Session'
-    ) -> ty.Optional[dict[str, str]]:
+    ) -> dict[str, str] | None:
         """Fetch authentication headers for message.
 
         :param session: The session object that the auth_plugin belongs to.
