@@ -29,19 +29,19 @@ class PasswordMethod(base.AuthMethod):
     """
 
     password: str
-    user_id: ty.Optional[str] = None
-    username: ty.Optional[str] = None
-    user_domain_id: ty.Optional[str] = None
-    user_domain_name: ty.Optional[str] = None
+    user_id: str | None = None
+    username: str | None = None
+    user_domain_id: str | None = None
+    user_domain_name: str | None = None
 
     def __init__(
         self,
         *,
         password: str,
-        user_id: ty.Optional[str] = None,
-        username: ty.Optional[str] = None,
-        user_domain_id: ty.Optional[str] = None,
-        user_domain_name: ty.Optional[str] = None,
+        user_id: str | None = None,
+        username: str | None = None,
+        user_domain_id: str | None = None,
+        user_domain_name: str | None = None,
     ) -> None:
         self.password = password
         self.user_id = user_id
@@ -55,7 +55,7 @@ class PasswordMethod(base.AuthMethod):
         auth: base.Auth,
         headers: dict[str, str],
         request_kwargs: dict[str, object],
-    ) -> ty.Union[tuple[None, None], tuple[str, ty.Mapping[str, object]]]:
+    ) -> tuple[None, None] | tuple[str, ty.Mapping[str, object]]:
         user: dict[str, ty.Any] = {'password': self.password}
 
         if self.user_id:
@@ -70,7 +70,7 @@ class PasswordMethod(base.AuthMethod):
 
         return 'password', {'user': user}
 
-    def get_cache_id_elements(self) -> dict[str, ty.Optional[str]]:
+    def get_cache_id_elements(self) -> dict[str, str | None]:
         return {
             'password_password': self.password,
             'password_user_id': self.user_id,
@@ -107,20 +107,20 @@ class Password(base.Auth):
         self,
         auth_url: str,
         password: str,
-        user_id: ty.Optional[str] = None,
-        username: ty.Optional[str] = None,
-        user_domain_id: ty.Optional[str] = None,
-        user_domain_name: ty.Optional[str] = None,
+        user_id: str | None = None,
+        username: str | None = None,
+        user_domain_id: str | None = None,
+        user_domain_name: str | None = None,
         *,
         unscoped: bool = False,
-        trust_id: ty.Optional[str] = None,
-        system_scope: ty.Optional[str] = None,
-        domain_id: ty.Optional[str] = None,
-        domain_name: ty.Optional[str] = None,
-        project_id: ty.Optional[str] = None,
-        project_name: ty.Optional[str] = None,
-        project_domain_id: ty.Optional[str] = None,
-        project_domain_name: ty.Optional[str] = None,
+        trust_id: str | None = None,
+        system_scope: str | None = None,
+        domain_id: str | None = None,
+        domain_name: str | None = None,
+        project_id: str | None = None,
+        project_name: str | None = None,
+        project_domain_id: str | None = None,
+        project_domain_name: str | None = None,
         reauthenticate: bool = True,
         include_catalog: bool = True,
     ) -> None:
