@@ -36,7 +36,7 @@ class ReceiptMethod(base.AuthMethod):
         auth: base.Auth,
         headers: dict[str, str],
         request_kwargs: dict[str, object],
-    ) -> ty.Union[tuple[None, None], tuple[str, ty.Mapping[str, object]]]:
+    ) -> tuple[None, None] | tuple[str, ty.Mapping[str, object]]:
         """Add the auth receipt to the headers.
 
         We explicitly return None to avoid being added to the request
@@ -45,5 +45,5 @@ class ReceiptMethod(base.AuthMethod):
         headers['Openstack-Auth-Receipt'] = self.receipt
         return (None, None)
 
-    def get_cache_id_elements(self) -> dict[str, ty.Optional[str]]:
+    def get_cache_id_elements(self) -> dict[str, str | None]:
         return {'receipt_receipt': self.receipt}
