@@ -30,19 +30,19 @@ class TOTPMethod(base.AuthMethod):
     """
 
     passcode: str
-    user_id: ty.Optional[str] = None
-    username: ty.Optional[str] = None
-    user_domain_id: ty.Optional[str] = None
-    user_domain_name: ty.Optional[str] = None
+    user_id: str | None = None
+    username: str | None = None
+    user_domain_id: str | None = None
+    user_domain_name: str | None = None
 
     def __init__(
         self,
         *,
         passcode: str,
-        user_id: ty.Optional[str] = None,
-        username: ty.Optional[str] = None,
-        user_domain_id: ty.Optional[str] = None,
-        user_domain_name: ty.Optional[str] = None,
+        user_id: str | None = None,
+        username: str | None = None,
+        user_domain_id: str | None = None,
+        user_domain_name: str | None = None,
     ) -> None:
         self.passcode = passcode
         self.user_id = user_id
@@ -56,7 +56,7 @@ class TOTPMethod(base.AuthMethod):
         auth: base.Auth,
         headers: dict[str, str],
         request_kwargs: dict[str, object],
-    ) -> ty.Union[tuple[None, None], tuple[str, ty.Mapping[str, object]]]:
+    ) -> tuple[None, None] | tuple[str, ty.Mapping[str, object]]:
         user: dict[str, ty.Any] = {'passcode': self.passcode}
 
         if self.user_id:
@@ -71,7 +71,7 @@ class TOTPMethod(base.AuthMethod):
 
         return 'totp', {'user': user}
 
-    def get_cache_id_elements(self) -> dict[str, ty.Optional[str]]:
+    def get_cache_id_elements(self) -> dict[str, str | None]:
         # NOTE(gyee): passcode is not static so we cannot use it as part of
         # the key in caching.
         return {
@@ -108,20 +108,20 @@ class TOTP(base.Auth):
         self,
         auth_url: str,
         passcode: str,
-        user_id: ty.Optional[str] = None,
-        username: ty.Optional[str] = None,
-        user_domain_id: ty.Optional[str] = None,
-        user_domain_name: ty.Optional[str] = None,
+        user_id: str | None = None,
+        username: str | None = None,
+        user_domain_id: str | None = None,
+        user_domain_name: str | None = None,
         *,
         unscoped: bool = False,
-        trust_id: ty.Optional[str] = None,
-        system_scope: ty.Optional[str] = None,
-        domain_id: ty.Optional[str] = None,
-        domain_name: ty.Optional[str] = None,
-        project_id: ty.Optional[str] = None,
-        project_name: ty.Optional[str] = None,
-        project_domain_id: ty.Optional[str] = None,
-        project_domain_name: ty.Optional[str] = None,
+        trust_id: str | None = None,
+        system_scope: str | None = None,
+        domain_id: str | None = None,
+        domain_name: str | None = None,
+        project_id: str | None = None,
+        project_name: str | None = None,
+        project_domain_id: str | None = None,
+        project_domain_name: str | None = None,
         reauthenticate: bool = True,
         include_catalog: bool = True,
     ) -> None:
